requirements.txt
mopidy
eyed3
beets
pillow
argparse
glob3
